refactor: Revert creation of DataFetch class. Instead make relevant methods in Issues.php static.
