+++ classes/BaseTab.php
@@ -21,7 +21,6 @@ abstract class BaseTab implements Tab {
   protected $displayShacl = false;
   protected $historicalDataDir = null;
   protected $versioning = false;
-  protected $version = null;
   protected $lang = 'en';
   public $analysisParameters = null;
   public $indexingParameters = null;
@@ -47,7 +46,7 @@ abstract class BaseTab implements Tab {
     $this->displayShacl = $this->configuration->doDisplayShacl(); // isset($configuration['display-shacl']) && (int) $configuration['display-shacl'] == 1;
     $this->versioning = $this->configuration->doVersioning(); // (isset($this->configuration['versions'][$this->db]) && $this->configuration['versions'][$this->db] === true);
 
-    $this->count = DataFetch::readCount($this->getFilePath('count.csv'));
+    $this->count = $this->readCount($this->getFilePath('count.csv'));
     $this->readLastUpdate();
     $this->handleHistoricalData();
     $this->lang = getOrDefault('lang', $this->catalogue->getDefaultLang(), ['en', 'de', 'pt']);
@@ -117,14 +116,21 @@ abstract class BaseTab implements Tab {
     return sprintf('%s/_historical/%s/%s/%s', $this->configuration->getDir(), $this->configuration->getDirName(), $version, $name);
   }
 
-  protected function filePath($filename) {
-    if ($this->versioning && $this->version != '') {
-      $elementsFile = $this->getVersionedFilePath($this->version, $filename);
+  protected function readCount($countFile = null) {
+    if (is_null($countFile))
+      $countFile = $this->getFilePath('count.csv');
+    if (file_exists($countFile)) {
+      $counts = readCsv($countFile);
+      if (empty($counts)) {
+        $count = trim(file_get_contents($countFile));
+      } else {
+        $counts = $counts[0];
+        $count = isset($counts->processed) ? $counts->processed : $counts->total;
+      }
     } else {
-      $elementsFile = $this->getFilePath($filename);
+      $count = 0;
     }
-    
-    return $elementsFile;
+    return intval($count);
   }
 
   protected function readLastUpdate() {
