+++ classes/Start.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once 'DataFetch.php';
+require_once 'Issues.php';
 
 class Start extends BaseTab {
 
@@ -11,6 +11,6 @@ class Start extends BaseTab {
   public function prepareData(Smarty &$smarty) {
     parent::prepareData($smarty);
 
-    $smarty->assign('issueStats', DataFetch::readTotal($this->filePath('issue-total.csv'), $this->count));
+    $smarty->assign('issueStats', Issues::readTotal($this->getFilePath('issue-total.csv'), $this->count));
   }
 }
